# schememitra/packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SchemeMitra Core API",
    "description": "Deterministic eligibility, repayment maths, partner routing, documents, voice and application tracking for marginalized entrepreneurs. Money is integer paise; rates are bps.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/voice/sessions": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Create Session",
        "description": "Start a guided voice intake. Works before sign-in, like the typed form.",
        "operationId": "create_session_v1_voice_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceTurnOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voice/sessions/{session_id}": {
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "Get Session State",
        "operationId": "get_session_state_v1_voice_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceTurnOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voice/sessions/{session_id}/utterance": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Utterance",
        "description": "One turn: audio (or typed text) → transcript → slots → localized reply (+ audio URL).",
        "operationId": "utterance_v1_voice_sessions__session_id__utterance_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_utterance_v1_voice_sessions__session_id__utterance_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceTurnOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voice/readback": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Readback",
        "description": "Summary of the application in the applicant's language, to be read aloud before submit (C6).",
        "operationId": "readback_v1_voice_readback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadbackOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voice/readback/{application_id}/reply": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Readback Reply",
        "description": "\"haan, sahi hai\" confirms (text evidence goes into the consent record); naming a field asks to correct it.",
        "operationId": "readback_reply_v1_voice_readback__application_id__reply_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_readback_reply_v1_voice_readback__application_id__reply_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadbackReplyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voice/tts/{digest}": {
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "Tts Clip",
        "operationId": "tts_clip_v1_voice_tts__digest__get",
        "parameters": [
          {
            "name": "digest",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{64}$",
              "title": "Digest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "audio/wav": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_readback_reply_v1_voice_readback__application_id__reply_post": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "audio": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio"
          }
        },
        "type": "object",
        "title": "Body_readback_reply_v1_voice_readback__application_id__reply_post"
      },
...
      "Body_utterance_v1_voice_sessions__session_id__utterance_post": {
        "properties": {
          "text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "audio": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio",
            "description": "≤15 s chunk: audio/webm (Opus), ogg, wav, mp4"
          }
        },
        "type": "object",
        "title": "Body_utterance_v1_voice_sessions__session_id__utterance_post"
      },
      "ClarifyOut": {
        "properties": {
          "slot": {
            "type": "string",
            "title": "Slot"
          },
          "value": {
            "title": "Value"
          },
          "display": {
            "type": "string",
            "title": "Display"
          }
        },
        "type": "object",
        "required": [
          "slot",
          "value",
          "display"
        ],
        "title": "ClarifyOut"
      },
...
      "DistrictHQ": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "title": "Lng"
          },
          "pincode": {
            "type": "string",
            "title": "Pincode"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng",
          "pincode"
        ],
        "title": "DistrictHQ",
        "description": "District headquarters, used as the search centre for nearby lenders (same as the typed form)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReadbackIn": {
        "properties": {
          "application_id": {
            "type": "string",
            "title": "Application Id"
          }
        },
        "type": "object",
        "required": [
          "application_id"
        ],
        "title": "ReadbackIn"
      },
      "ReadbackOut": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Url"
          },
          "tts": {
            "type": "string",
            "enum": [
              "clip",
              "bhashini",
              "device"
            ],
            "title": "Tts",
            "default": "device"
          },
          "demo_reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo Reply"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ReadbackOut"
      },
      "ReadbackReplyOut": {
        "properties": {
          "intent": {
            "type": "string",
            "enum": [
              "confirm",
              "correct",
              "unclear"
            ],
            "title": "Intent"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "transcript": {
            "type": "string",
            "title": "Transcript"
          },
          "transcript_engine": {
            "type": "string",
            "enum": [
              "bhashini",
              "whisper",
              "sandbox",
              "typed"
            ],
            "title": "Transcript Engine"
          },
          "sandbox": {
            "type": "boolean",
            "title": "Sandbox",
            "default": false
          },
          "reply_text": {
            "type": "string",
            "title": "Reply Text"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "intent",
          "transcript",
          "transcript_engine",
          "reply_text"
        ],
        "title": "ReadbackReplyOut"
      },
...
      "SessionIn": {
        "properties": {
          "lang": {
            "type": "string",
            "maxLength": 8,
            "title": "Lang",
            "description": "one of en, hi, bn, mr, te, ta, gu, ur, kn, or, ml, pa, as",
            "default": "en"
          }
        },
        "type": "object",
        "title": "SessionIn"
      },
...
      "SlotOut": {
        "properties": {
          "value": {
            "title": "Value"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "display": {
            "type": "string",
            "title": "Display"
          }
        },
        "type": "object",
        "required": [
          "value",
          "confidence",
          "method",
          "display"
        ],
        "title": "SlotOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoiceTurnOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "lang": {
            "type": "string",
            "title": "Lang"
          },
          "state": {
            "type": "string",
            "enum": [
              "intake",
              "clarify",
              "confirm",
              "done"
            ],
            "title": "State"
          },
          "turn": {
            "type": "integer",
            "title": "Turn"
          },
          "max_turns": {
            "type": "integer",
            "title": "Max Turns"
          },
          "transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript"
          },
          "transcript_engine": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "bhashini",
                  "whisper",
                  "sandbox",
                  "typed"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript Engine"
          },
          "sandbox": {
            "type": "boolean",
            "title": "Sandbox",
            "default": false
          },
          "reply_text": {
            "type": "string",
            "title": "Reply Text"
          },
          "reply_audio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply Audio Url"
          },
          "tts": {
            "type": "string",
            "enum": [
              "clip",
              "bhashini",
              "device"
            ],
            "title": "Tts",
            "default": "device"
          },
          "slots": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SlotOut"
            },
            "type": "object",
            "title": "Slots"
          },
          "asked": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Asked"
          },
          "changed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changed",
            "default": []
          },
          "clarify": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClarifyOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "handoff": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "rules",
                  "form"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Handoff"
          },
          "district_hq": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DistrictHQ"
              },
              {
                "type": "null"
              }
            ]
          },
          "filled": {
            "type": "integer",
            "title": "Filled"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "elapsed_ms": {
            "type": "integer",
            "title": "Elapsed Ms"
          },
          "processing_ms": {
            "type": "integer",
            "title": "Processing Ms"
          },
          "demo_utterances": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Demo Utterances",
            "default": []
          },
          "asr_available": {
            "type": "boolean",
            "title": "Asr Available",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "lang",
          "state",
          "turn",
          "max_turns",
          "reply_text",
          "slots",
          "asked",
          "filled",
          "total",
          "elapsed_ms",
          "processing_ms"
        ],
        "title": "VoiceTurnOut"
      }
    }
  }
}